requirements.txt
pillow==9.5.0
matplotlib==3.7.1